+++ tests/round.test.js
@@ -1,4 +1,5 @@
 import { Decimal128 } from "../src/decimal128.mjs";
+import * as string_decoder from "string_decoder";
 
 const roundingModes = [
     "ceil",
@@ -50,7 +51,7 @@ describe("rounding", () => {
         });
         test("more digits than are available", () => {
             expect(new Decimal128("1.5").round(2).toString()).toStrictEqual(
-                "1.5"
+                "1.50"
             );
         });
         test("negative odd", () => {
@@ -60,9 +61,23 @@ describe("rounding", () => {
         });
         test("round down (positive)", () => {
             expect(new Decimal128("1.1").round(6).toString()).toStrictEqual(
-                "1.1"
+                "1.100000"
             );
         });
+    });
+    test("integer", () => {
+        expect(new Decimal128("42").round().toString()).toStrictEqual("42");
+    });
+    test("negative integer", () => {
+        expect(new Decimal128("-42").round().toString()).toStrictEqual("-42");
+    });
+    test("negative number of digits requested", () => {
+        expect(() => new Decimal128("1.5").round(-42)).toThrowError(RangeError);
+    });
+    test("too many digits requested", () => {
+        expect(() => new Decimal128("1.5").round(2 ** 53)).toThrowError(
+            RangeError
+        );
     });
 });